Add example 'Dummy' driver.
The example Dummy driver is just a static driver that takes some
parameters and puts them in a structure. That's pretty much it.
It could be simpler if you wanted something more bare, but the
intention is that just the c/fan file, and the fan configuration
in c/fans be changed to make an implementation that communicates
with real hardware.