Add a 'managed' control mode.
In discussions with Chris, it's been obvious that we expect to have
a module that would be able to control the fans on the system - an
overall manager for the system's fans and other sensors. Such a
system isn't really under manual control, but is being managed.

This change introduces a managed mode which sits alongside the manual
mode. The driver is oblivious to this mode - to it, this is just
manual. This means that there's a few changes in the way that the
control is set up - we now have to remember the control mode. This
also means that the controller won't call down to the driver as
often because we now keep the state.

Documentation has been updated to include more description of the
different control modes and how they relate.