Fix fan_setcontrol issuing driver calls on Managed -> Manual changes.
The transition from Managed->Manual would trigger a call to the fan
driver, even though the fan driver didn't actually care - it was only
going to see the Manual state. It's not a huge difference, but as I've
taken care to not call the driver unnecessarily, it makes sense to
trim that case too.