Add some example code; document the location id bitfield.
A couple of example BASIC programs are included to demonstrate the
use of a couple of the SWIs. These are intended just to give an easy
entry to trying the system out, and not to be comprehensive at
present.

In writing these, it was obvious that the documentation was lacking -
the fan location was never actually documented as a bitfield, only
the values were listed. These are now properly documented.