Add CI building using the RISC OS build service.
The build was using the cross-compiling toolchain. That's not so nice,
as I am the only person who has this. So a GitHub CI and GitLab CI
configuration has been created, together with a suitable robuild.yaml
to build on the RISC OS build service.

This should mean that once it's published out to public GitHub it will
be able to be tested there too!

The gitlab build still uses the cross-compiling build because... well,
whilst it's internal we want to test that!