Add trap for the service call for fan changed, to pass to pollword.
The FanDriver is able to report the state change to error, if it
detects one. This is because the fan driver would usually receive
this notification out of band (ie without the FanController being
involved). Whilst it might be possible to pass the state change to
FanController to notify clients, that's a little redundant, so the
driver just raises this as a notification itself and FanController
takes note.

We note this by setting any pollwords that requested it.