Apply registered SWI number, errors and services.
The SWI number, error base and service base have now been registered.
This means that the definitions have changed in the header file,
and the documentation has been updated to reflect the new numbers.
The API version has also been incremented to 1.01 as strictly the API
has changed - although anyone using 1.00 wouldn't be able to get the
information from the 1.01 system because the SWI interface has changed.